# docs/swagger.yaml
definitions:
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.AccountResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      photos:
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      role:
        $ref: '#/definitions/models.AccountRole'
      username:
        type: string
    type: object
  models.AccountRole:
    enum:
    - admin
    - normal
    type: string
    x-enum-varnames:
    - ROLE_ADMIN
    - ROLE_NORMAL
  models.Comment:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      message:
        type: string
      photo_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.CreateAccount:
    properties:
      age:
        type: integer
      email:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/models.AccountRole'
      username:
        type: string
    required:
    - age
    - email
    - password
    - role
    - username
    type: object
  models.LoginAccount:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.Socialmedia:
    properties:
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: string
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
  socialmedia.SocialMediaCreate:
    properties:
      name:
        type: string
      social_media_url:
        type: string
      user_id:
        type: string
    required:
    - name
    - social_media_url
    type: object
  token.Tokens:
    properties:
      access_token:
        type: string
      id_token:
        type: string
      refresh_token:
        type: string
    type: object
info:
  contact: {}
paths:
  /accounts:
    get:
      description: get an user account
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AccountResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: get user account
      tags:
      - User
    post:
      consumes:
      - application/json
      description: create user by inputing the correct user datas
      parameters:
      - description: Create Account Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateAccount'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.AccountResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: create user
      tags:
      - User
  /accounts/login:
    post:
      consumes:
      - application/json
      description: fetch user information by id
      parameters:
      - description: Login Account Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.LoginAccount'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/token.Tokens'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: finding user record
      tags:
      - User
...
  /socialmedias:
    post:
      consumes:
      - application/json
      description: Creates a new social media with the provided data
      parameters:
      - description: Create Social Media Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/socialmedia.SocialMediaCreate'
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Socialmedia'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new social media
      tags:
      - Social Media
  /socialmedias/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a social media with the given id
      parameters:
      - description: Social Media ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a social media by ID
      tags:
      - Social Media
    get:
      consumes:
      - application/json
      description: Fetch a social media with the given id
      parameters:
      - description: SocialMedia ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Socialmedia'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Find a social media by ID
      tags:
      - Social Media
    put:
      consumes:
      - application/json
      description: Updates an existing social media with the provided data
      parameters:
      - description: SocialMedia ID
        in: path
        name: id
        required: true
        type: string
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Create Social Media Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Socialmedia'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update an existing social media by id
      tags:
      - Social Media
  /socialmedias/all:
    get:
      description: fetch all social media records
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Socialmedia'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: finds all social media records
      tags:
      - Social Media
swagger: "2.0"